requirements.txt
# 기본 의존성
numpy>=1.20.0
torch>=1.9.0
torchaudio>=0.9.0
transformers>=4.11.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
hydra-core>=1.1.0
python-dotenv>=0.19.0
langdetect>=1.0.9

# 차원 축소 및 시각화
umap-learn>=0.5.3
numba>=0.55.0  # UMAP 의존성 
langdetect>=1.0.9
